refactor(#20): shared/utils 폴더 import 경로 변경
