+++ shared/utils/index.ts
@@ -1,1 +1,4 @@
 export { classMerge } from './classMerge'
+export { useGotoTop } from './useGotoTop'
+export { useToast } from './useToast'
+export { useDebounce } from './useDebounce'
